add controlled tooltip
